requirements.txt
streamlit
python-dotenv
pypdf
google-generativeai
json
time
os
io
